+++ crates/voicevox_core/src/engine/song/queries.rs
@@ -29,7 +29,17 @@ pub use self::{key::Key, optional_lyric::OptionalLyric};
 ///
 /// # Validation
 ///
-/// この構造体の状態によっては、`Synthesizer`の各メソッドは[`ErrorKind::InvalidQuery`]を表わすエラーを返す。詳細は[`validate`メソッド]にて。
+/// この構造体は不正な状態を表現しうる。どのような状態が不正なのかについては[`validate`メソッド]を参照。この構造体を使う関数は、不正な状態に対して[`ErrorKind::InvalidQuery`]を表わすエラーを返す。
+///
+/// [`Deserialize`]時には、不正な状態であるかの検証は行われない。外部からのデータが不正でないことを確かめるには、デシリアライズ後に`validate`メソッドを用いる必要がある。
+///
+/// ```
+/// # use voicevox_core::Score;
+/// # let json = r#"{ "notes": [{ "lyric": "", "frame_length": 0 }] }"#;
+/// let score = serde_json::from_str::<Score>(json)?;
+/// score.validate()?;
+/// # anyhow::Ok(())
+/// ```
 ///
 /// [`ErrorKind::InvalidQuery`]: crate::ErrorKind::InvalidQuery
 /// [`validate`メソッド]: Self::validate
@@ -50,7 +60,17 @@ impl From<&'_ Score> for serde_json::Value {
 ///
 /// # Validation
 ///
-/// この構造体の状態によっては、`Synthesizer`の各メソッドは[`ErrorKind::InvalidQuery`]を表わすエラーを返す。詳細は[`validate`メソッド]にて。
+/// この構造体は不正な状態を表現しうる。どのような状態が不正なのかについては[`validate`メソッド]を参照。この構造体を使う関数は、不正な状態に対して[`ErrorKind::InvalidQuery`]を表わすエラーを返す。
+///
+/// [`Deserialize`]時には、不正な状態であるかの検証は行われない。外部からのデータが不正でないことを確かめるには、デシリアライズ後に`validate`メソッドを用いる必要がある。
+///
+/// ```
+/// # use voicevox_core::Note;
+/// # let json = r#"{ "lyric": "", "frame_length": 0 }"#;
+/// let note = serde_json::from_str::<Note>(json)?;
+/// note.validate()?;
+/// # anyhow::Ok(())
+/// ```
 ///
 /// [`ErrorKind::InvalidQuery`]: crate::ErrorKind::InvalidQuery
 /// [`validate`メソッド]: Self::validate
@@ -106,6 +126,15 @@ macro_rules! key {
 
 impl Key {
     const NAME: &str = "音階";
+
+    const fn saturating_from(n: u8) -> Self {
+        assert!(Self::MIN.get() == 0);
+        if let Some(ret) = Self::__new(n) {
+            ret
+        } else {
+            Self::MAX
+        }
+    }
 }
 
 impl FromStr for Key {
@@ -436,9 +465,13 @@ mod key {
         /// 最大値。`127`。
         pub const MAX: Self = Self(127);
 
+        /// ビット長。
+        pub const BITS: u32 = u8::BITS - Key::MAX.0.leading_zeros();
+
         #[doc(hidden)]
         pub const fn __new(n: u8) -> Option<Self> {
-            if Self::MIN.get() <= n && n <= Self::MAX.get() {
+            const _: () = assert!(Key::MIN.0 == 0);
+            if n <= Self::MAX.0 {
                 Some(Self(n))
             } else {
                 None
@@ -448,6 +481,170 @@ mod key {
         pub const fn get(self) -> u8 {
             self.0
         }
+
+        /// [`{integer}::strict_add`]と同じことを行う。
+        ///
+        /// # Panics
+        ///
+        /// ## Overflow behavior
+        ///
+        /// オーバーフローするならパニックする。
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use voicevox_core::key;
+        ///
+        /// assert_eq!(key!(61), key!(60).strict_add(1));
+        /// ```
+        ///
+        /// ```should_panic
+        /// # use voicevox_core::Key;
+        /// #
+        /// Key::MAX.strict_add(1);
+        /// ```
+        ///
+        /// [`{integer}::strict_add`]: u8::strict_add
+        #[must_use = "same reason as `{integer}::strict_add`"]
+        #[track_caller]
+        pub const fn strict_add(self, rhs: u8) -> Self {
+            if let Some(n) = self.0.checked_add(rhs) {
+                if let Some(ret) = Self::__new(n) {
+                    return ret;
+                }
+            }
+            panic!("attempt to add with overflow");
+        }
+
+        /// [`u{n}::strict_add_signed`]と同じことを行う。
+        ///
+        /// # Panics
+        ///
+        /// ## Overflow behavior
+        ///
+        /// オーバーフローするならパニックする。
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use voicevox_core::key;
+        ///
+        /// assert_eq!(key!(59), key!(60).strict_add_signed(-1));
+        /// ```
+        ///
+        /// ```should_panic
+        /// # use voicevox_core::Key;
+        /// #
+        /// Key::MAX.strict_add_signed(1);
+        /// ```
+        ///
+        /// ```should_panic
+        /// # use voicevox_core::key;
+        /// #
+        /// key!(0).strict_add_signed(-1);
+        /// ```
+        ///
+        /// [`u{n}::strict_add_signed`]: u8::strict_add_signed
+        #[must_use = "same reason as `u{n}::strict_add_signed`"]
+        #[track_caller]
+        pub const fn strict_add_signed(self, rhs: i8) -> Self {
+            if let Some(n) = self.0.checked_add_signed(rhs) {
+                if let Some(ret) = Self::__new(n) {
+                    return ret;
+                }
+            }
+            panic!("attempt to add with overflow");
+        }
+
+        /// [`{integer}::strict_sub`]と同じことを行う。
+        ///
+        /// # Panics
+        ///
+        /// ## Overflow behavior
+        ///
+        /// オーバーフローするならパニックする。
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use voicevox_core::key;
+        ///
+        /// assert_eq!(key!(59), key!(60).strict_sub(1));
+        /// ```
+        ///
+        /// ```should_panic
+        /// # use voicevox_core::key;
+        /// #
+        /// key!(0).strict_sub(1);
+        /// ```
+        ///
+        /// [`{integer}::strict_sub`]: u8::strict_sub
+        #[must_use = "same reason as `{integer}::strict_sub`"]
+        #[track_caller]
+        pub const fn strict_sub(self, rhs: u8) -> Self {
+            if let Some(n) = self.0.checked_sub(rhs) {
+                if let Some(ret) = Self::__new(n) {
+                    return ret;
+                }
+            }
+            panic!("attempt to subtract with overflow");
+        }
+
+        // TODO: Rust 1.90以降になったら`strict_sub_signed`も追加する。
+
+        /// [`{integer}::saturating_add`]と同じことを行う。
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use voicevox_core::key;
+        ///
+        /// assert_eq!(key!(61), key!(60).saturating_add(1));
+        /// assert_eq!(key!(127), key!(126).saturating_add(2));
+        /// ```
+        ///
+        /// [`{integer}::saturating_add`]: u8::saturating_add
+        #[must_use = "same reason as `{integer}::saturating_add`"]
+        pub const fn saturating_add(self, rhs: u8) -> Self {
+            Self::saturating_from(self.0.saturating_add(rhs))
+        }
+
+        /// [`u{n}::saturating_add_signed`]と同じことを行う。
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use voicevox_core::key;
+        ///
+        /// assert_eq!(key!(61), key!(60).saturating_add_signed(1));
+        /// assert_eq!(key!(0), key!(1).saturating_add_signed(-2));
+        /// assert_eq!(key!(127), key!(126).saturating_add_signed(2));
+        /// ```
+        ///
+        /// [`u{n}::saturating_add_signed`]: u8::saturating_add_signed
+        #[must_use = "same reason as `u{n}::saturating_add_signed`"]
+        pub const fn saturating_add_signed(self, rhs: i8) -> Self {
+            Self::saturating_from(self.0.saturating_add_signed(rhs))
+        }
+
+        /// [`{integer}::saturating_sub`]と同じことを行う。
+        ///
+        /// # Examples
+        ///
+        /// ```
+        /// use voicevox_core::key;
+        ///
+        /// assert_eq!(key!(59), key!(60).saturating_sub(1));
+        /// assert_eq!(key!(0), key!(1).saturating_sub(2));
+        /// ```
+        ///
+        /// [`{integer}::saturating_sub`]: u8::saturating_sub
+        #[must_use = "same reason as `{integer}::saturating_sub`"]
+        pub const fn saturating_sub(self, rhs: u8) -> Self {
+            Self::__new(self.0.saturating_sub(rhs)).expect("should be in the range")
+        }
+
+        // TODO: Rust 1.90以降になったら`saturating_sub_signed`も追加する。
     }
 }
 
@@ -462,7 +659,7 @@ mod optional_lyric {
         mora_mappings::MORA_KANA_TO_MORA_PHONEMES,
     };
 
-    /// 音符の歌詞。`""`は[無音]。
+    /// 音符の歌詞。空文字列は[無音]。
     ///
     /// # Examples
     ///
@@ -554,6 +751,9 @@ mod optional_lyric {
 
 #[cfg(test)]
 mod tests {
+    use std::cmp;
+
+    use itertools::iproduct;
     use rstest::rstest;
 
     use super::key::Key;
@@ -567,6 +767,78 @@ mod tests {
                 MIN..=MAX => assert!(Key::__new(n).is_some()),
                 n => assert!(Key::__new(n).is_none()),
             }
+        }
+    }
+
+    #[test]
+    fn key_strict_add_works() {
+        for (lhs, rhs) in iproduct!(Key::MIN.get()..=Key::MAX.get(), 0..=u8::MAX) {
+            let lhs = Key::__new(lhs).unwrap();
+            // TODO: Rust 2024にしたらlet chainにする
+            if let Some(sum) = lhs.get().checked_add(rhs) {
+                if sum <= Key::MAX.get() {
+                    assert_eq!(sum, lhs.strict_add(rhs).get());
+                }
+            }
+        }
+    }
+
+    #[test]
+    fn key_strict_add_signed_works() {
+        for (lhs, rhs) in iproduct!(Key::MIN.get()..=Key::MAX.get(), i8::MIN..=i8::MAX) {
+            let lhs = Key::__new(lhs).unwrap();
+            // TODO: Rust 2024にしたらlet chainにする
+            if let Some(sum) = lhs.get().checked_add_signed(rhs) {
+                if sum <= Key::MAX.get() {
+                    assert_eq!(sum, lhs.strict_add_signed(rhs).get());
+                }
+            }
+        }
+    }
+
+    #[test]
+    fn key_strict_sub_works() {
+        for (lhs, rhs) in iproduct!(Key::MIN.get()..=Key::MAX.get(), 0..=u8::MAX) {
+            let lhs = Key::__new(lhs).unwrap();
+            // TODO: Rust 2024にしたらlet chainにする
+            if let Some(sum) = lhs.get().checked_sub(rhs) {
+                if sum <= Key::MAX.get() {
+                    assert_eq!(sum, lhs.strict_sub(rhs).get());
+                }
+            }
+        }
+    }
+
+    #[test]
+    fn key_saturating_add_works() {
+        for (lhs, rhs) in iproduct!(Key::MIN.get()..=Key::MAX.get(), 0..=u8::MAX) {
+            let lhs = Key::__new(lhs).unwrap();
+            assert_eq!(
+                cmp::min(lhs.get().saturating_add(rhs), Key::MAX.get()),
+                lhs.saturating_add(rhs).get(),
+            );
+        }
+    }
+
+    #[test]
+    fn key_saturating_add_signed_works() {
+        for (lhs, rhs) in iproduct!(Key::MIN.get()..=Key::MAX.get(), i8::MIN..=i8::MAX) {
+            let lhs = Key::__new(lhs).unwrap();
+            assert_eq!(
+                cmp::min(lhs.get().saturating_add_signed(rhs), Key::MAX.get()),
+                lhs.saturating_add_signed(rhs).get(),
+            );
+        }
+    }
+
+    #[test]
+    fn key_saturating_sub_works() {
+        for (lhs, rhs) in iproduct!(Key::MIN.get()..=Key::MAX.get(), 0..=u8::MAX) {
+            let lhs = Key::__new(lhs).unwrap();
+            assert_eq!(
+                cmp::min(lhs.get().saturating_sub(rhs), Key::MAX.get()),
+                lhs.saturating_sub(rhs).get(),
+            );
         }
     }