+++ CHANGELOG.md
@@ -53,8 +53,10 @@
 
 ### Added
 
-- \[Rust,Python\] ソング機能が追加されます ([#531], [#732], [#738], [#761], [#895], [#896], [#894], [#1217], [#1236], [#1073], [#1242], [#1250], [#1244])。
-- \[Python,Java\] 一部のドキュメントの文体が改善されます ([#1238])。
+- \[Rust,Python\] ソング機能が追加されます ([#531], [#732], [#738], [#761], [#895], [#896], [#894], [#1217], [#1236], [#1073], [#1242], [#1250], [#1252], [#1247], [#1253], [#1244])。
+- ドキュメントが改善されます。
+    - [バージョン0.16.3](#0163---2025-12-08-0900)で導入された、`AudioQuery`/`AccentPhrase`/`Mora`のバリデーション機能に関するドキュメンテーションがよりわかりやすくなります ([#1251])。
+    - \[Python,Java\] 一部のドキュメントの文体が改善されます ([#1238])。
 
 ### Changed
 
@@ -1404,7 +1406,11 @@ Windows版ダウンローダーのビルドに失敗しています。
 [#1238]: https://github.com/VOICEVOX/voicevox_core/pull/1238
 [#1242]: https://github.com/VOICEVOX/voicevox_core/pull/1242
 [#1244]: https://github.com/VOICEVOX/voicevox_core/pull/1244
+[#1247]: https://github.com/VOICEVOX/voicevox_core/pull/1247
 [#1250]: https://github.com/VOICEVOX/voicevox_core/pull/1250
+[#1251]: https://github.com/VOICEVOX/voicevox_core/pull/1251
+[#1252]: https://github.com/VOICEVOX/voicevox_core/pull/1252
+[#1253]: https://github.com/VOICEVOX/voicevox_core/pull/1253
 
 [VOICEVOX/onnxruntime-builder#25]: https://github.com/VOICEVOX/onnxruntime-builder/pull/25